Add files via upload
This branch from early April 2018 resolved find and replace issues between the Municipal Court Civil and Common Pleas Court Civil subroutines of the program. These issues arose when the seven original programs were merged under a single GUI as both of the subroutines called a function with the same name that performed different tasks. This consolidates find and replace tasks into a single findreplace() function that meets the needs of both subroutines. It also eliminates the use of regular expressions in the findreplace() function and now calls the built in string function .replace().